+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+# v1.0.4
+## 08/16/2017
+
+1. [](#new)
+    * CLI: Added `status` command to check config and git (#52, thanks @karfau)
+    * Allow local branches to be named differently than the remote branches (#48, thanks @denniswebb)
+    * Added support for new Admin Navigation Tray
+1. [](#bugfix)
+    * Fixed minimum Git required version to support `--all` (#32,#49, thanks @redrohX)
+
 # v1.0.3
 ## 02/21/2017